A lot of Stuff
We need to add some opening and closing game buttons and just update the main screen. Other than that, the game works for each of the 9 squares a new set of nine squares show up. It knows when you win because I've kept track of the position in the array list. I'll comment the code later to explain whats goin on.