Added private field mainButtons
This private variable gives you access to the main buttons on the screen. Also go ahead and download this: http://i.imgur.com/BSp08yW.png  This is the main x button. I've only programmed for x's to show on the main menu if won. For o's I just have a place holder, "o's won" until i work on designing the O's for the main game. I haven't kept track of the main games exes and o's but that'll be easy considering we have access to those buttons now.